backend/requirements.txt
Flask
Flask-Cors
Pillow
numpy
opencv-python-headless
torch
torchvision
transformers
scikit-learn
numpy-stl
pyvista
trimesh
gunicorn
foamlib
openai
llama-index-core
jinja2